feat(cli): add opt-in blz-dev profile tooling
